refactor/RSS-ECOMM-2_01: replace validation tests to a separate file
